rename module name ssim->circuit_simulation
