+++ bluecellulab/__init__.py
@@ -18,7 +18,7 @@ from .psection import PSection  # NOQA
 from .psegment import PSegment  # NOQA
 from .simulation import Simulation  # NOQA
 from .rngsettings import RNGSettings  # NOQA
-from .ssim import CircuitSimulation, CircuitSimulation  # NOQA
+from .circuit_simulation import CircuitSimulation, CircuitSimulation  # NOQA
 
 
 from .simulation.neuron_globals import NeuronGlobals
